R-Car-Gen3: update energy model
The capacity value for A53 cpu (379) is wrong.
The dhrystone score has been measured, and the capacity value depends on it.

Our previous capacity model is below. (R-Car H3)

                 dhrystone   capacity
  ------------------------------------
  CA57@500MHz   3040[DMIPS]     308
  CA57@1000MHz  5894[DMIPS]     596
  CA57@1500MHz  8879[DMIPS]     898   default setting
  CA57@1600MHz  9563[DMIPS]     965   turbo
  CA57@1700MHz 10120[DMIPS]    1024   turbo
  CA53@1200MHz  3745[DMIPS]     379
  ------------------------------------

However, it seems to not support the cases of turbo mode in current
capacity feature. So I except turbo case, and I define new capacity
mode below.

                 dhrystone   capacity
  ------------------------------------
  CA57@500MHz   3040[DMIPS]     351
  CA57@1000MHz  5894[DMIPS]     680
  CA57@1500MHz  8879[DMIPS]    1024   default setting
  CA53@1200MHz  3745[DMIPS]     432
  ------------------------------------

So the following parameters are correct.

  cpu_capacity@CA57 = 1024
  cpu_capacity@CA53 =  432

However it also seems to not support the cases of big.little in
current capacity feature. Actually, the capacity parameters must be
set below.

  capacity-dmips-mhz@CA57 = 1024
  capacity-dmips-mhz@CA53 =  432/1200*1500 = 540

Signed-off-by: Frode Isaksen <fisaksen@baylibre.com>